run_analysis.R
Please submit a link to a Github repo with the code for performing your
analysis. The code should have a file run_analysis.R in the main
directory that can be run as long as the Samsung data is in your working
directory. The output should be the tidy data set you submitted for part
1. You should include a README.md in the repo describing how the script
works and the code book describing the variables.